backend/requirements.txt
fastapi
uvicorn
redis
google-generativeai
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]